Extract query hooks
