+++ features/issues/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-get-issues";
+export * from "./use-resolve-issue";
