+++ features/issues/index.ts
@@ -1,2 +1,3 @@
+export * from "./api";
 export * from "./components/issue-list";
 export * from "./types/issue.types";
