+++ features/projects/index.ts
@@ -1,3 +1,3 @@
-export * from "./api/use-projects";
+export * from "./api";
 export * from "./components/project-list";
 export * from "./types/project.types";
